Seeder 練習

+++ database/migrations/2020_12_10_082553_create_students_table.php
@@ -15,7 +15,7 @@ class CreateStudentsTable extends Migration
     {
         Schema::create('students', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('user_id')->nullable();
+            $table->bigInteger('user_id')->nullable();
             $table->string('no')->unique();
             $table->string('tel');
             $table->timestamps();
